pyproject.toml
[tool.poetry]
name = "event-reporter"
version = "0.1.0"
description = ""
authors = ["Kirill Lapushinskiy <kirilllapushinskiy@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
